requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bertopic==0.15.0
bidict==0.22.1
blis==0.7.11
build==0.10.0
catalogue==2.0.10
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
coloredlogs==15.0.1
confection==0.1.3
constantly==15.1.0
contourpy==1.2.0
cryptography==41.0.2
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.8
Cython==0.29.36
decorator==5.1.1
distro==1.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0.tar.gz#sha256=7ef2a0090b49aaab02d6eba347186e3d4ff99328334f5504e1da3afe2b3474e0
Events==0.4
exceptiongroup==1.2.0
feedparser==6.0.10
filelock==3.12.2
flake8==6.1.0
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.9.2
geographiclib==2.0
geopy==2.4.0
h11==0.14.0
hdbscan==0.8.33
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.16.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langcodes==3.3.0
llvmlite==0.41.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.1
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
netifaces==0.10.6
netkit==3.1.12
networkx==3.1
nltk==3.8.1
numba==0.58.1
numpy==1.25.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
openai==1.3.5
packaging==23.1
pandas==2.0.3
parsel==1.8.1
pathy==0.10.2
Pillow==10.1.0
plotly==5.18.0
preshed==3.0.9
Protego==0.2.1
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyDispatcher==2.0.7
pyflakes==3.1.0
pynndescent==0.5.10
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2023.3
PyYAML==6.0.1
queuelib==1.6.2
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
rpds-py==0.9.2
safetensors==0.3.3
scikit-learn==1.3.2
scipy==1.11.3
Scrapy==2.9.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==23.1.0
setproctitle==1.3.2
sgmllib3k==1.0.0
simple-websocket==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
socketio==0.2.1
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
tbb==2021.10.0
tenacity==8.2.3
thinc==8.1.12
threadpoolctl==3.2.0
tldextract==3.4.4
tokenizers==0.14.0
tomli==2.0.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.65.0
transformers==4.34.0
triton==2.1.0
Twisted==22.10.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
umap-learn==0.5.4
urllib3==2.0.3
validators==0.20.0
w3lib==2.1.1
wasabi==1.1.2
websocket-client==1.6.4
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.0
